Basic support vor relative widths
When the "width" property of the css style for ul.token-input-list
is relative (e.g. "width: 100%") or omitted, the width of the
drop down list is adjusted to be as wide as the token list itself.
(The "width" for div.token-input-dropdown is overwritten.)

As there is no "resize" event for boxes, the implementation is
rather clumsy, and already visible drop downs are not resized
(immediately) when the token list box changes its size.