Preliminary Support for lazy rendering of the results list
When the search result consists of a lot of items, it may take the
web browser a long time to render all of them at once.  This patch
introduces lazy rendering of the results.  This means, only the
visible part plus a buffer above and below is rendered.  When the
User scrolls beyond the rendered buffer, a new part of the results
is rendered.

Currently, the lazy rendering has only been tested using locally
generated data.  One of the next commits introduces the also
necessary counterpart "lazy loading".