Fix issues w/ "Lazy Rendering" in Internet Explorer
The rendering itself already worked in IE, but it was not possible
to scroll the dropdown box by dragging the scroll bar handle with
the mouse:  When clicking the scollbar, the input field lost its
focus.  This subsequently cancelled the selecting process by hiding
the dropdown field.

The problem has been fixed by guessing why the input field has lost
its focus and giving it back when the reason appears to be scrolling.

Tested with IE 7, 8, 9 and 10 and FF 3.6.22 and 5.0.1 and Opera 11.5.