Re-Calaculate the drop down box' z-index when needed (again)
vitorbal's solution w/ wrapping the dropdown into a div that
uses relative positioning does not work when the widget is
placed in a jQuery UI Modal.  The Dropdown cannot expand
beyond the modal's bounding box anymore.

The Dropdown is now attached to the document 'body' again and a
Z-Index is calculated.