Prevent the dropdown from expanding beyond the browser viewport.
The dropdown's height is still limited using the css attribute
'max-height'.

This addition prevents the dropdown from only being partially
visible because it would go below or above the browser's current
viewport.

In case it would go beyond the lower border of the browser,
and there is more space above the token box, the dropdown is
placed above the box.  Regardless of whether is rendered
below or above the token box, the vertical size is limited so
that it remains completely visible.