Allow hintText to be a watermark in the input box
Add option to allow hintText to be a watermark inside the input
box instead of appearing on the dropdown.
When this option is enabled a new class is also added to the input
element to help customize the text styling of the watermark.
This patch also adds a demo of the new functionality to the demo page.

Fixes issue #203 from upstream