Replace current rendering strategy w/ smarter one.
The current rendering stategy for lazy rendering was to replace
all rendered dropdown items when the currently rendered area was
left (by scrolling up or down).

The new rendering strategy only adds and removes items depending
on whether they are in close proximity to the visible area.

The change is very necessary, because the current rendering strategy
makes the dropdown area to jump up and down when scrolling.